address PR review feedback
- Switch dev-only auto-email-confirm to #if DEBUG 2so it cannot ship to Release - Make review aggregate update race-safe via Serializable tx + recompute from persisted rows - Extract GetCurrentProfileAsync into ReviewsController.Functions.cs - Guard against int overflow in pagination Skip - Reset reviews page when ProfileLoaded gets a new profile.id - Fix failing pagination test: use IsAssignableFrom for BadRequestObjectResult; add overflow test - Split ReviewDialog / SubmitCompletionButton / ApproveCompletionButton / ApplicationControls / TaskApplicationsPanel into their own files
