+++ backend/Features/Auth/AuthController.cs
@@ -27,7 +27,6 @@ public sealed partial class AuthController(
     IClientIpAccessor clientIpAccessor,
     IEmailSender emailSender,
     IOptions<EmailOptions> emailOptions,
-    IHostEnvironment environment,
     ILogger<AuthController> logger) : ControllerBase
 {
     private const string RefreshTokenCookieName = "refreshToken";
@@ -133,13 +132,14 @@ public sealed partial class AuthController(
         await db.SaveChangesAsync(cancellationToken);
         await transaction.CommitAsync(cancellationToken);
 
-        if (environment.IsDevelopment())
-        {
-            var confirmToken = await userManager.GenerateEmailConfirmationTokenAsync(user);
-            await userManager.ConfirmEmailAsync(user, confirmToken);
-            return Ok(new RegisterResponse("Registration successful. Your account is ready to use."));
-        }
-
+#if DEBUG
+        // Auto-confirm email in local Debug builds so seeded dev users can log in
+        // without a working SendGrid key. Compile-time guard so this can never
+        // reach Release output.
+        var confirmToken = await userManager.GenerateEmailConfirmationTokenAsync(user);
+        await userManager.ConfirmEmailAsync(user, confirmToken);
+        return Ok(new RegisterResponse("Registration successful. Your account is ready to use."));
+#else
         bool emailSent = true;
         try
         {
@@ -156,6 +156,7 @@ public sealed partial class AuthController(
             : "Registration successful, but we could not send the verification email. Use the resend-verification endpoint to request a new link.";
 
         return Ok(new RegisterResponse(message));
+#endif
     }
 
     [HttpPost("login")]
